dev environments

+++ cogs/cog_list/essentials/basic.py
@@ -29,13 +29,16 @@ async def update_guild_list_database():
             firebase.edit_data(endpoint_path, {guild.id: default_json})
 
 
-    bot_guild_ids = [guild.id for guild in global_vars.global_bot.guilds]
-    for guild_id in firebase.get_data(endpoint_path):
-        guild_id = int(guild_id)
-        if guild_id not in bot_guild_ids:
-            firebase.delete_data(
-                firebase.ShortEndpoint.discord_guilds.get_path() + [guild_id]
-            )
+    if not global_vars.dev_environment:
+        bot_guild_ids = [guild.id for guild in global_vars.global_bot.guilds]
+        for guild_id in firebase.get_data(endpoint_path):
+            guild_id = int(guild_id)
+            if guild_id not in bot_guild_ids:
+                firebase.delete_data(
+                    firebase.ShortEndpoint.discord_guilds.get_path() + [guild_id]
+                )
+    else:
+        pass
 
 
 class CogEvents(cog.RegisteredCog):
@@ -62,7 +65,12 @@ class CogEvents(cog.RegisteredCog):
     @cmds.Cog.listener()
     async def on_guild_remove(self, guild: nx.Guild):
         """On guild leave."""
-        await update_guild_list_database()
+        if not global_vars.dev_environment:
+            await update_guild_list_database()
+        else:
+            firebase.delete_data(
+                firebase.ShortEndpoint.discord_guilds.get_path() + [guild.id]
+            )
 
 
     @disc_utils.command_wrap(
